der_write_tags exec spec; monadic serialize specs [skip ci]

+++ src/Lib/Types.v
@@ -58,9 +58,9 @@ Inductive Err := HeaderEncodeError
                  | CustomError {T : Type} : T -> Err.
 
 (* Specialized version of errW with custom Error and Log type with specialized tell *)
-Definition errW1 := @errW Err (list byte).
+Definition errW1 := @errW Err (list int).
 
 Existing Class Monoid.
 Existing Instance Monoid_list_app.
 
-Definition tell1 := @tell (list byte) (Monoid_list_app) errW1 Writer_errW.
+Definition tell1 := @tell (list int) (Monoid_list_app) errW1 Writer_errW.
